refactor: introduce PacketProcessorHolder interface and allow user register their own packet processor
