+++ Allay-Server/src/main/java/org/allaymc/server/network/processor/login/RequestNetworkSettingsPacketProcessor.java
@@ -1,6 +1,7 @@
 package org.allaymc.server.network.processor.login;
 
 import org.allaymc.api.entity.interfaces.EntityPlayer;
+import org.allaymc.api.network.processor.login.ILoginPacketProcessor;
 import org.allaymc.api.server.Server;
 import org.cloudburstmc.protocol.bedrock.packet.BedrockPacketType;
 import org.cloudburstmc.protocol.bedrock.packet.NetworkSettingsPacket;
